* psantos@lsd.di.uminho.pt--pson/neem--nat--0.5--base-0
   tag of psantos@lsd.di.uminho.pt--pson/neem--work--0.4--patch-2

 * psantos@lsd.di.uminho.pt--pson/neem--nat--0.5--patch-1
   Merge from neem--work--0.4--patch-7

 * psantos@lsd.di.uminho.pt--pson/neem--nat--0.5--patch-2
   Fiddling with membership values and timings. Nothing big.

 * psantos@lsd.di.uminho.pt--pson/neem--nat--0.5--patch-3
   Removed a println

 * psantos@lsd.di.uminho.pt--pson/neem--nat--0.5--patch-4
   Code cleaning. Removed a java.util.Properties testing class and removed dependency from printUtils package in MembershipImpl class.

 * psantos@lsd.di.uminho.pt--pson/neem--work--0.4--base-0
   tag of jop@di.uminho.pt--prj/neem--head--0.4--patch-4

 * psantos@lsd.di.uminho.pt--pson/neem--work--0.4--patch-1
   Minor cleaning

 * psantos@lsd.di.uminho.pt--pson/neem--work--0.4--patch-2
   minor changes

 * psantos@lsd.di.uminho.pt--pson/neem--work--0.4--patch-3
   Minor changes to try and solve the membership's incoherence.

 * psantos@lsd.di.uminho.pt--pson/neem--work--0.4--patch-4
   Memberhip layer not gossiping peer anymore, now just sends to one peer. Changes have also been introduced in other files in order to try to solve UUID associated problems ( The dynamic membership is not dynamic at all. Hope to get it right soon.

 * psantos@lsd.di.uminho.pt--pson/neem--work--0.4--patch-5
   nothing to report

 * psantos@lsd.di.uminho.pt--pson/neem--work--0.4--patch-6
   Dynamic membership problem solved. Membership changes now visible very quickly. If the local group size increases, the number of connected peers rises. If the local group size is reduced, connections are closed very quickly to ensure the new local group's size. There was no problem with the MembershipImpl class. The problem was in the testing application wich binded the socket to one address and not to the INADDR_ANY.

 * psantos@lsd.di.uminho.pt--pson/neem--work--0.4--patch-7
   AbstractGossipImpl no longer needed. GossipImpl inherited its methods and variables.

+++ src/net/sf/neem/impl/Bucket.java
@@ -57,7 +57,7 @@ import java.nio.*;
 public class Bucket {
     
     /** Creates a new instance of Bucket */
-    public Bucket(ByteBuffer[] msg, Integer port) {
+    public Bucket(ByteBuffer[] msg, Short port) {
         this.msg = msg;
         this.port = port;
     }
@@ -68,12 +68,12 @@ public class Bucket {
     }
     
     /** Returns the port to wich the message is intended */
-    public Integer getPort() {
+    public Short getPort() {
         return (this.port);
     }
     
     private ByteBuffer[] msg;
-    private Integer port;
+    private Short port;
 }
 
 // arch-tag: b07a5de4-0eca-4ba3-9bf5-c1564cd9cece
